Refaktorointi done?
