+++ application/linkkivinkit/src/main/java/linkkivinkki/app/command/Command.java
@@ -1,6 +1,8 @@
 
 package linkkivinkki.app.command;
 
+import java.util.HashMap;
+
 public interface Command {
     boolean execute();
 }
